Migrate task to ts
