+++ src/pages/Content/modules/icons/discussion.tsx
@@ -1,0 +1,23 @@
+import React from 'react';
+
+const DiscussionIcon = (
+  <svg
+    style={{
+      fill: 'white',
+      height: '25px',
+      width: '25px',
+    }}
+    version="1.1"
+    viewBox="0 0 1920 1920"
+    xmlns="http://www.w3.org/2000/svg"
+  >
+    <path
+      d="M677.647059,16 L677.647059,354.936471 L790.588235,354.936471 L790.588235,129.054118 L1807.05882,129.054118 L1807.05882,919.529412 L1581.06353,919.529412 L1581.06353,1179.29412 L1321.41176,919.529412 L1242.24,919.529412 L1242.24,467.877647 L677.647059,467.877647 L0,467.877647 L0,1484.34824 L338.710588,1484.34824 L338.710588,1903.24706 L756.705882,1484.34824 L1242.24,1484.34824 L1242.24,1032.47059 L1274.99294,1032.47059 L1694.11765,1451.59529 L1694.11765,1032.47059 L1920,1032.47059 L1920,16 L677.647059,16 Z M338.789647,919.563294 L903.495529,919.563294 L903.495529,806.622118 L338.789647,806.622118 L338.789647,919.563294 Z M338.789647,1145.44565 L677.726118,1145.44565 L677.726118,1032.39153 L338.789647,1032.39153 L338.789647,1145.44565 Z M112.941176,580.705882 L1129.41176,580.705882 L1129.41176,1371.40706 L710.4,1371.40706 L451.651765,1631.05882 L451.651765,1371.40706 L112.941176,1371.40706 L112.941176,580.705882 Z"
+      fillRule="evenodd"
+      stroke="none"
+      strokeWidth="1"
+    />
+  </svg>
+);
+
+export default DiscussionIcon;
