+++ src/pages/Content/modules/icons/index.ts
@@ -1,0 +1,5 @@
+import AssignmentIcon from './assignment';
+import DiscussionIcon from './discussion';
+import QuizIcon from './quiz';
+
+export { AssignmentIcon, DiscussionIcon, QuizIcon };
