+++ src/pages/Content/modules/types/course.ts
@@ -1,6 +1,8 @@
-export interface Course {
+interface Course {
   id: number;
   color: string;
   name: string;
   position: number;
 }
+
+export default Course;
