+++ src/pages/Content/modules/types/index.ts
@@ -1,3 +1,4 @@
-import { Course } from './course';
+import Course from './course';
+import Assignment from './assignment';
 
-export { Course };
+export { Course, Assignment };
